==> Leena/02_calculate_vessel_lengths_Leena.Rmd <==
---
title: "Calculate Vessel Lengths for DCRB Vessels"
output:  html_document
---

## Purpose

Calculate vessel lengths based on vessel registration data, and then add in a few columns to the fish ticket data providing vessel length for each ticket.

We use a decision tree for calculating vessel length from PacFIN data. The template allows for one year or multiple years of fish tickets to be run at the same time. To run the vessel length calculations for *all* vessels in the landings data, set option `dcrb_only` to `FALSE` in Setup and Options (this is the default).

Currently, the code writes out processed vessel length keys and combined vessel length/fish ticket landings data as `.rds` files. Code to write to `.csv` maintained but commented out.

Plotting of outputs is optional, indicated

## Methods Overview:

* Step 1: Pull the registration data up to two years prior to the fish ticket year (3 years total)

* Step 2: Remove any registration entries with vessels larger than 200 feet and smaller than 10 feet

* Step 3: Find the number of unique vessel lengths for all unique Vessel Number / Agency Code combinations.

* Step 4: Calculate final vessel length for all unique Vessel Number / Agency Code combinations
- If vessel length data was available for 2+ years and MAX_LENGTH < MIN_LENGTH + 10, then assume this reflects an actual increase or decrease in vessel size.Final vessel length is mean of the two most recent vessel lengths
- If vessel length data was available for only one year OR vessel length data was available for 2+ years, but MAX_LENGTH > MIN_LENGTH + 10. Pull registration data up to four years prior to the fish ticket year (5 years total). Then, if...

  a. One vessel length was recorded in 2+ years. Use this value as vessel length.
  b. Two different vessel lengths were recorded.
    i. MAX_LENGTH > 2 x MIN_LENGTH. <-- this is probably a decimal point error that we would need to check manually to determine true length. Denote NA.
    ii. MAX_LENGTH < 2 X MIN_LENGTH. <-- Use the mode of these lengths
  c. Three or more different vessel lengths were recorded. Use the median of these lengths

### Script Includes:

1. Calculation of vessel lengths for X years of fish tickets.

2. Summary of calculation output: proportion of fishing vessels missing lengths (per year), distribution of vessel lengths (per year and overall), vessel length against the number of years vessel length was recorded (per year), and the frequency of the calculation types used (per year and overall).

3. Application of calculated vessel lengths to appropriate fish tickets.

Resources:

[PMSC Fleet Report](https://www.psmfc.org/efin/docs/fleetreport.pdf), esp Table 8
[PacFIN Column names key](https://pacfin.psmfc.org//wp-content/uploads/2016/06/PacFIN_Comprehensive_Fish_Tickets.pdf)

***

## Setup and Options

Load packages
```{r}
library(tidyverse)
library(here)
library(magrittr)
library(knitr)

# ggplot theme
plot_theme <-   theme_minimal()+
  theme(text=element_text(family="sans",size=12,color="black"),
        legend.text = element_text(size=14),
        axis.title=element_text(family="sans",size=14,color="black"),
        axis.text=element_text(family="sans",size=8,color="black"),
        panel.grid.major = element_line(color="gray50",linetype=3))
theme_set(plot_theme)


rm(list=setdiff(ls(),c('process_year','alltime')))

```

Source functions for vessel length calculations.
```{r source_functions}
source(here::here('process steps',"Report_Vessel_Length_processedtix_functions.R"))
```

Enter processing options: Year desired to be processed, whether to only process Dungeness vessels, and whether to plot outputs.

```{r process options}
# process_year <- 2020
dcrb_only <- FALSE
plot_output <- TRUE
```

## Vessel Permit Data

Load permits data

```{r load perm}

#2020-2021 data differs from other years:
permitsfiles_2020_2021 <- list.files(here::here('data','raw','vessel registration 2020_2021' ),full.names = T)

permits_2020_2021 <- purrr::map_df(permitsfiles_2020_2021,function(fl){
  read_csv(fl,col_types='ddcddccdcccccdddccccccddllcl')
}) %>% distinct()

#remove column that doesn't appear in earlier years, and move one column to mathc earlier years
permits_2020_2021 <- permits_2020_2021 %>% 
  select(-IS_USED_FOR_CORRECTION) %>% 
  relocate(VESSEL_OWNER_PHONE_NUMBER, .after = VESSEL_OWNER_ADDRESS_ZIP)



#other years
permitsfiles <- list.files(here::here('data','raw','vessel registration'),full.names = T)

permits <- purrr::map_df(permitsfiles,function(fl){
  read_csv(fl,col_types='ddcddccdcccccdddcccccccddll')
}) %>% distinct()


# join old and new permits
permits <- rbind(permits,permits_2020_2021)
```

## Subset Permit Data

Choose columns of interest and subset the data frame
```{r pthin}
pcols <- c("VESSEL_NUM", "AGENCY_CODE","REGISTRATION_YEAR","VESSEL_LENGTH", "LENGTH_TYPE_CODE")
pthin <- permits %>%  select(pcols)
```

Change all entries of `777` for vessel length to `NA` (based on recommendation from Blake Feist)

```{r}
pthin %<>%
  mutate(VESSEL_LENGTH=na_if(VESSEL_LENGTH,777))
```

Remove any vessel lengths greater than 200 or lower than 10. *Note - depending on the fishery, may want to change these cutoffs*

```{r length_filter}
pthin_length_filter <- pthin %>%
  filter(VESSEL_LENGTH < 200,VESSEL_LENGTH > 10)
```

## Vessels of Interest

Read in raw fish ticket `.rds` files for each year of interest to data frame "landings". If option `dcrb_only` above is `TRUE`, filter to only vessels with positive revenue for Dungeness crab in the landings data.

```{r}
# import landings data
#Adjust import file name - use one that has FTID columns
landings <- purrr::map(process_year, function(y){
  read_rds(here::here('data','processed','fish tickets',paste0(y,'fishtix_withFTID.rds')))
}) %>% 
  bind_rows()

# find relevant vessels (all vessels versus Dungeness vessels only)
vessels <- landings %>%
  filter(drvid != "UNKNOWN", drvid != "MISSING", drvid != "", !is.na(drvid)) %>%
  select(drvid, agency_code, year) %>%
  distinct()

if(dcrb_only){
  vessels <- landings %>%
    filter(drvid != "UNKNOWN", drvid != "MISSING", drvid != "", !is.na(drvid)) %>%
    filter(DCRB_revenue > 0) %>%
    select(drvid, agency_code, year) %>%
    distinct()
}
```

## Calculate Vessel Lengths

Initiate empty data frame for all vessel length data across years
```{r}
vessel_length_key_df <- tibble("drvid" = character(),
                               "agency_code" = character(),
                               "year" = numeric(),
                               "FINAL_LENGTH" = numeric(),
                               "TYPE_CALC" = character(),
                               "UNIQUE_LENGTHS" = numeric(),
                               "N_YEARS_LENGTH_RECORDED" = numeric(),
                               "HISTORIC_DATA" = character())
```

Do the calculation of vessel lengths.

```{r calc_lengths_multiyr}
for(y in unique(vessels$year)){
  # take only the vessels fishing in year "y"
  year_vessels <- filter(vessels, year == y)
  
  # identify the years of regulation data to pull
  target_reg_years <- seq(y-2, y)
  
  # subset for target registration years and summarise permits for each vessel
  cat("Calculating 3yr summary statistics for vessels in ", y, "\n")
  pthin_sumstats <- pthin_length_filter %>%
    filter(REGISTRATION_YEAR %in% target_reg_years,!is.na(VESSEL_LENGTH)) %>%
    group_by(VESSEL_NUM, AGENCY_CODE) %>%
    arrange(desc(REGISTRATION_YEAR)) %>%
    summarise(n_lengths = length(VESSEL_LENGTH),
              n_unique = length(unique(VESSEL_LENGTH)),
              max_length = max(VESSEL_LENGTH),
              min_length = min(VESSEL_LENGTH),
              mean2yr = get2yrmean(x=VESSEL_LENGTH, years=REGISTRATION_YEAR)) %>% 
    ungroup()
  
  # create empty vectors for this year
  final_vessel_lengths <- c()
  length_calc_vec <- c()
  n_unique_vec <- c()
  n_lengths_vec <- c()
  historic_vec <- c()
  processed <- 0
  
  cat("Calculating vessel lengths for fishing vessels in ", y, "...\n")
  # for each vessel fishing in this year #
  for(i in seq(1:length(year_vessels$drvid))){
    ## use the calc_length function (loaded from the "functions.R" file) to calculate vessel length
    tmp_vessel_length_info <- calc_length(permits=permits, vesseldat = year_vessels, lengthdat = pthin_length_filter, summarydat = pthin_sumstats, index = i)
    ## save the calc_length output to the appropriate position ("i") in the output vectors for this year
    n_lengths_vec[i] <- tmp_vessel_length_info[1] %>% as.numeric()
    n_unique_vec[i] <- tmp_vessel_length_info[2] %>% as.numeric()
    final_vessel_lengths[i] <- tmp_vessel_length_info[3] %>% as.numeric()
    length_calc_vec[i] <- tmp_vessel_length_info[4]
    ## if the vessel had to be calculated with historical data from over 5 years ago, a warning message will be saved in the calc_length output
    if(length(tmp_vessel_length_info) > 4){
      ### print the warning message
      print(tmp_vessel_length_info[5])
      ### save "Y" to the historic_vec for this year
      historic_vec[i] <- "Y"
    } else{ historic_vec[i] <- "N" }
    processed <- processed + 1
  }
  cat("Done processing", processed, "vessels for", y, "\n")
  # save allof the output vectors to a data frame for this year
  tmp_vessel_length_key_df <- tibble("drvid" = year_vessels$drvid,
                                     "agency_code" = year_vessels$agency_code,
                                     "year" = year_vessels$year,
                                     "FINAL_LENGTH" = final_vessel_lengths,
                                     "TYPE_CALC" = length_calc_vec,
                                     "UNIQUE_LENGTHS" = n_unique_vec,
                                     "N_YEARS_LENGTH_RECORDED" = n_lengths_vec,
                                     "HISTORIC_DATA" = historic_vec)
  ## bind this year's data frame to the end of the full data frame
  vessel_length_key_df <- bind_rows(vessel_length_key_df, tmp_vessel_length_key_df)
  cat("Wrote out", dim(tmp_vessel_length_key_df)[1], "lengths for", y, " to final data frame\n\n")
}
```

## Explore Vessel Lengths

How many fishing vessels WITH Vessel Numbers are missing calculated lengths?

```{r}
cat(sum(is.na(vessel_length_key_df$FINAL_LENGTH)), 'vessels with Vessel ID numbers are missing calculated lengths, or', sum(is.na(vessel_length_key_df$FINAL_LENGTH)) / length(vessel_length_key_df$FINAL_LENGTH)*100, 'percent.')
```
```{r}
for(y in unique(vessel_length_key_df$year)){
  for(a in unique(vessel_length_key_df$agency_code)){
    tmp_dat <- vessel_length_key_df %>%
      filter(agency_code == a) %>%
      filter(year == y)
    missing <- sum(is.na(tmp_dat$FINAL_LENGTH))
    cat("Number", y, "Vessels Missing Vessel Lengths for", a, ":", missing, "\n")
    cat("Proportion:", missing/length(tmp_dat$FINAL_LENGTH), "\n\n")
  }
}
```

```{r}
# ggplot(data=filter(vessel_length_key_df, !is.na(FINAL_LENGTH)), aes(x=as.numeric(FINAL_LENGTH),y=as.numeric(N_YEARS_LENGTH_RECORDED))) +
#   geom_point()+
#   facet_wrap(~year) +
#   xlab("Calculated Vessel Length") +
#   ylab("Number of Years Length Was Recorded") +
#   theme(axis.text.x = element_blank())
# Frequency table of number of years' records
vessel_length_key_df %>% 
  count(N_YEARS_LENGTH_RECORDED) %>% 
  kable(col.names = c("Number of Records","Number of Vessels"))
```

If plotting option above is `TRUE`, make some informative plots.
```{r length freq}
if(plot_output){
  vessel_length_key_df %>% 
    filter(!is.na(FINAL_LENGTH)) %>% 
    ggplot(aes(as.numeric(FINAL_LENGTH))) +
      geom_histogram(bins=10) +
      facet_wrap(~year) +
      labs(x="Vessel Length",y="Number of Vessels",title="Vessel Length Frequency Distribution")
}
```

```{r calc types}
if(plot_output){
  vessel_length_key_df %>% 
    ggplot(aes(vessel_length_key_df$TYPE_CALC)) +
    geom_bar() +
    facet_wrap(~year) +
    labs(x="Calculation Type",y="Number of Vessels")+
    theme(axis.text.x = element_text(angle=90, hjust=1))
}

```

## Assess Missing Data for Fish Tickets

For each year, join vessel length data with landings. How many fish tickets are missing a vessel length, by agency? Non-specific.
```{r}
for(y in unique(vessel_length_key_df$year)){
  landings_subset <- landings %>%
    filter(year == y)
  dcrb_landings_length <- left_join(landings_subset, filter(vessel_length_key_df, year == y), by=c("drvid", "agency_code"))
  for(a in unique(dcrb_landings_length$agency_code)){
    tmp_dat <- filter(dcrb_landings_length, agency_code == a)
    missing <- sum(is.na(tmp_dat$FINAL_LENGTH))
    cat("Number of", y, "fish tickets missing vessel lengths for", a, ":", missing, "\n")
    cat("Proportion:", missing/length(tmp_dat$FINAL_LENGTH), "\n\n")
  }
}
```

## Save Results

Vessel lengths key

```{r eval=TRUE}
for(y in unique(vessel_length_key_df$year)){
  key_subset <- vessel_length_key_df %>%
    filter(year == y)
  write_rds(key_subset,here::here('data','processed','vessel length keys',paste0("vessel_length_key_", y, ".rds")))
}
```

Landings matched to vessel lengths

```{r eval=TRUE}
#save output with diff name to indicate FTID still in data
for(y in unique(vessel_length_key_df$year)){
  landings_subset <- landings %>%
    filter(year == y)
  vessel_length_key_df_subset <- vessel_length_key_df %>%
    filter(year==y) %>%
    select(-year)
  dcrb_landings_length <- left_join(landings_subset,vessel_length_key_df_subset, by=c("drvid", "agency_code"))
  write_rds(dcrb_landings_length,here::here('data','processed','fish tickets',paste0(y, "fishtix_vlengths_withFTID.rds")))
}
# for(y in unique(vessel_length_key_df$year)){
#   landings_subset <- landings %>%
#     filter(year == y)
#   vessel_length_key_df_subset <- vessel_length_key_df %>%
#     filter(year==y) %>%
#     select(-year)
#   dcrb_landings_length <- left_join(landings_subset,vessel_length_key_df_subset, by=c("drvid", "agency_code"))
#   write.csv(file= paste0(fishtixdir, "processed_fishtickets_with_vlength_", y, ".csv"), x = dcrb_landings_length, row.names = FALSE)
# }
```
